% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df500}
\alias{df500}
\title{a sample of 500 records from smk_lungca.der7612_deadff in the BigQuery test database}
\format{data.frame}
\source{
NHS
}
\usage{
df500
}
\description{
a sample of 500 records from smk_lungca.der7612_deadff in the BigQuery test database
}
\examples{
names(bq4nhs::df500)
}
\keyword{datasets}
