% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhsBQ.R
\name{nhsBQ}
\alias{nhsBQ}
\title{generate a BigQueryConnection to a selected project/dataset}
\usage{
nhsBQ(project = "test-project1-205019", ds = "der7612_deadff")
}
\arguments{
\item{project}{character(1) project name in BigQuery}

\item{ds}{character(1) dataset name}
}
\description{
generate a BigQueryConnection to a selected project/dataset
}
\examples{
nhsBQ()
}
