% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdef.R
\name{survdef}
\alias{survdef}
\title{produce a query that generates fields life_months and death_observed
based on NHS irt76, age76, and deathmonth}
\usage{
survdef(bqconn, tablename, extras = NULL)
}
\arguments{
\item{bqconn}{a bigrquery BigQueryConnection instance, or a tibble}

\item{tablename}{character(1) a table in the connection dataset}

\item{extras}{defaults to NULL; or a vector of character() variables to be
retrieved from the source table}
}
\value{
a tibble if input is a tibble
}
\description{
produce a query that generates fields life_months and death_observed
based on NHS irt76, age76, and deathmonth
}
\note{
The algorithm for computing life months is
'''
life_months = ifelse(is.na(deathmonth), (age76*12)+(maxintake-912),
     deathmonth+(age76*12)-912)  # 912 is 76 years, a temporal anchor for study beginning in 1976
'''
}
\examples{
survdef(tibble::as.tibble(bq4nhs::df500), extras=c("smkdr80", "bmi80"))
}
